Separate events from travels. Now events are displayed in dedicated view not in modal form. Add edit event functionality

+++ src/main/java/com/example/TravelPlanner/travel/TravelRepository.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 @Repository
 public interface TravelRepository extends JpaRepository<Travel, Long> {
 
-    Optional<Travel> findById(int id);
+    Optional<Travel> findById(Long id);
 
     List<Travel> findByAppUserId(Long appUserId);
